# api/Controllers/RoadController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

using api.Dtos;
using api.Dtos.Roads;
using api.Extensions;
using api.Interfaces;
using api.Models;

using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

using Swashbuckle.AspNetCore.Annotations;

namespace api.Controllers
{
    [ApiController]
    [Route("api/road")]
    public class RoadController : ControllerBase
    {
        private readonly IRoadService _roadService;
        public RoadController(IRoadService roadService)
        {
            _roadService = roadService;
        }

        [ProducesResponseType(typeof(List<RoadDto>), StatusCodes.Status200OK)]
        [HttpGet]
        [AllowAnonymous]
        [SwaggerOperation("Get all roads")]
        public async Task<IActionResult> GetAllRoads()
        {
            var roads = await _roadService.GetAllRoads();
            return Ok(roads);
        }

        [ProducesResponseType(typeof(RoadDto), StatusCodes.Status200OK)]
        [ProducesResponseType(typeof(ResponseModel), StatusCodes.Status400BadRequest)]
        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
        [HttpPost]
        [Authorize]
        [SwaggerOperation("Create a new road")]
        public async Task<IActionResult> CreateRoad([FromBody] CreateRoadDto dto)
        {
            if (!User.IsAccessToken()) return Unauthorized();
            if (!ModelState.IsValid) return BadRequest(ModelState);

            var road = await _roadService.CreateRoad(dto, User.GetId());
            return Ok(road);
        }

        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(typeof(ResponseModel), StatusCodes.Status404NotFound)]
        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
        [HttpPost("rating")]
        [Authorize]
        [SwaggerOperation("Add rating to road")]
        public async Task<IActionResult> AddRatingToSpot([FromBody] CreateRatingDto dto)
        {
            if (!User.IsAccessToken()) return Unauthorized();
            if (!ModelState.IsValid) return BadRequest(ModelState);

            var response = await _roadService.RateRoad(dto, User.GetId());
            if (response == null) return Ok();
            if (response.Status == "Error: NotFound") return NotFound(response);
            return BadRequest(response);
        }
    }
}